requirements.txt
flask==3.0.3
gunicorn==22.0.0
python-dotenv==1.0.1
py-clob-client>=0.22.0
web3>=7.0.0
